client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "MCP gRPC Client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio",
    "grpcio-tools",
    "python-dotenv",
    "httpx"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
